refactor(cli): add TODOs + minor refactors
